requirements.txt
sentence-transformers
torch
fastapi
uvicorn
pydantic
numpy
qdrant-client
openai
python-dotenv
